SPT-1514: Move EVCS related code to new `credential-store` directory
A pure refactor that starts to bring together code that is related to
interacting with EVCS into a new directory. This includes code that
calls EVCS extracted from handlers and types that represent an EVCS
response. Some types have been renamed. This aims to help developers
understand the codebase and make it easier to build upon.
