+++ src/credential-store/credential-store-error-response.ts
