+++ src/credential-store/credential-store-identity-response.ts
@@ -1,0 +1,16 @@
+export type CredentialStoreIdentityResponse = {
+  si: JWTIncludingStateAndMetadata;
+  vcs: JWTIncludingStateAndMetadata[];
+  afterKey?: string;
+};
+
+interface JWTIncludingStateAndMetadata {
+  state: string;
+  vc: string;
+  metadata: Metadata | null | string;
+  signature?: string;
+}
+
+interface Metadata {
+  [key: string]: unknown;
+}
